+++ include/votca/csg/cgengine.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2009-2019 The VOTCA Development Team (http://www.votca.org)
+ * Copyright 2009-2021 The VOTCA Development Team (http://www.votca.org)
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,12 +15,13 @@
  *
  */
 
-#ifndef _VOTCA_CSG_CGENGINE_H
-#define _VOTCA_CSG_CGENGINE_H
+#ifndef VOTCA_CSG_CGENGINE_H
+#define VOTCA_CSG_CGENGINE_H
 
 // Standard includes
 #include <list>
 #include <map>
+#include <memory>
 
 // Third party includes
 #include <boost/program_options.hpp>
@@ -56,12 +57,11 @@ namespace csg {
 class CGEngine {
  public:
   CGEngine();
-  ~CGEngine();
 
   /**
       create a coarse grained topolgy based on a given topology
   */
-  TopologyMap *CreateCGTopology(Topology &in, Topology &out);
+  std::unique_ptr<TopologyMap> CreateCGTopology(Topology &in, Topology &out);
 
   /**
       load molecule type from file
@@ -84,24 +84,24 @@ class CGEngine {
   bool IsIgnored(std::string ident);
 
  private:
-  std::map<std::string, CGMoleculeDef *> _molecule_defs;
+  std::map<std::string, std::unique_ptr<CGMoleculeDef>> _molecule_defs;
 
   std::list<std::string> _ignores;
 };
 
 inline CGMoleculeDef *CGEngine::getMoleculeDef(std::string name) {
-  std::map<std::string, CGMoleculeDef *>::iterator iter;
+  std::map<std::string, std::unique_ptr<CGMoleculeDef>>::iterator iter;
 
   // if there is only 1 molecule definition, don't care about the name
   if (_molecule_defs.size() == 1 && name == "unnamed") {
-    return (*(_molecule_defs.begin())).second;
+    return (*(_molecule_defs.begin())).second.get();
   }
 
   iter = _molecule_defs.find(name);
   if (iter == _molecule_defs.end()) {
     return nullptr;
   }
-  return (*iter).second;
+  return (*iter).second.get();
 }
 
 inline bool CGEngine::IsIgnored(std::string ident) {
@@ -116,4 +116,4 @@ inline bool CGEngine::IsIgnored(std::string ident) {
 }  // namespace csg
 }  // namespace votca
 
-#endif /* _VOTCA_CSG_CGENGINE_H */
+#endif  // VOTCA_CSG_CGENGINE_H